Move verifyCluster to utils.go
